# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: MCP Server for coding workflows — file editing, code execution, git, package management, and more
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /files/write:
    post:
      operationId: writeFile
      tags: [files]
      summary: Write or update a file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileWrite"
      responses:
        "200":
          description: File written
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileEntry"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /execute:
    post:
      operationId: executeCode
      tags: [execution]
      summary: Execute code in a sandboxed environment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteRequest"
      responses:
        "200":
          description: Execution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /execute/sessions/{sessionId}:
    delete:
      operationId: killExecutionSession
      tags: [execution]
      summary: Kill an execution session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session killed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---- Terminal ----
...
  /env/{key}:
    get:
      operationId: getEnvVar
      tags: [environment]
      summary: Get a specific environment variable
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Environment variable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvVarEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: setEnvVar
      tags: [environment]
      summary: Set an environment variable
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnvVarInput"
      responses:
        "200":
          description: Environment variable set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvVarEntry"
    delete:
      operationId: deleteEnvVar
      tags: [environment]
      summary: Delete an environment variable
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Variable deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---- Auth ----
  /auth/keys:
    get:
      operationId: listApiKeys
      tags: [auth]
      summary: List API keys
      responses:
        "200":
          description: API key list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKey"
    post:
      operationId: createApiKey
      tags: [auth]
      summary: Create a new API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeyInput"
      responses:
        "201":
          description: API key created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyWithSecret"
  /auth/keys/{keyId}:
    delete:
      operationId: revokeApiKey
      tags: [auth]
      summary: Revoke an API key
      parameters:
        - name: keyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Key revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Key not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---- Logs ----
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: ["string", "null"]

    # Files
    FileEntry:
      type: object
      required: [name, path, type, size, modifiedAt]
      properties:
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [file, directory, symlink]
        size:
          type: integer
        modifiedAt:
          type: string
        permissions:
          type: ["string", "null"]
...
    FileWrite:
      type: object
      required: [path, content]
      properties:
        path:
          type: string
        content:
          type: string
        encoding:
          type: string
          default: utf-8
        createDirs:
          type: boolean
          default: true
...
    ExecuteRequest:
      type: object
      required: [language, code]
      properties:
        language:
          type: string
          enum: [node, python, bash]
        code:
          type: string
        timeout:
          type: integer
          default: 30
        workDir:
          type: ["string", "null"]
        env:
          type: ["object", "null"]
          additionalProperties:
            type: string
        args:
          type: array
          items:
            type: string
    ExecutionResult:
      type: object
      required: [stdout, stderr, exitCode, duration]
      properties:
        stdout:
          type: string
        stderr:
          type: string
        exitCode:
          type: integer
        duration:
          type: number
        sessionId:
          type: ["string", "null"]
        timedOut:
          type: boolean
...
    EnvVarEntry:
      type: object
      required: [key, redacted]
      properties:
        key:
          type: string
        value:
          type: ["string", "null"]
        redacted:
          type: boolean
        group:
          type: ["string", "null"]
    EnvVarInput:
      type: object
      required: [value]
      properties:
        value:
          type: string
        redacted:
          type: boolean
          default: false

    # Auth
    ApiKey:
      type: object
      required: [id, name, prefix, createdAt, lastUsedAt, permissions]
      properties:
        id:
          type: string
        name:
          type: string
        prefix:
          type: string
        createdAt:
          type: string
        lastUsedAt:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
        permissions:
          type: array
          items:
            type: string
        requestCount:
          type: integer
    ApiKeyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        permissions:
          type: array
          items:
            type: string
        expiresAt:
          type: ["string", "null"]
    ApiKeyWithSecret:
      type: object
      required: [id, name, prefix, key, createdAt, permissions]
      properties:
        id:
          type: string
        name:
          type: string
        prefix:
          type: string
        key:
          type: string
        createdAt:
          type: string
        expiresAt:
          type: ["string", "null"]
        permissions:
          type: array
          items:
            type: string

    # Logs
...