# lib/api-zod/src/generated/types/executeRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MCP Server for coding workflows — file editing, code execution, git, package management, and more
 * OpenAPI spec version: 1.0.0
 */
import type { ExecuteRequestEnv } from './executeRequestEnv';
import type { ExecuteRequestLanguage } from './executeRequestLanguage';

export interface ExecuteRequest {
  language: ExecuteRequestLanguage;
  code: string;
  timeout?: number;
  /** @nullable */
  workDir?: string | null;
  /** @nullable */
  env?: ExecuteRequestEnv;
  args?: string[];
}